Adding exception testing to cover invalid instructions
This covers illegal instructions (like traps, that are not implemented)
as well as instructions that throw errors (like unaligned memops).

Implemented using a kernel module that installs an exception handler and
reports the exceptions to the user app (and automatically skips the
problematic instruction). On unexpected exceptions the upper/previous
handler will be invoked, which makes it work well with psplink.
Only single stepping on the "bad instruction" fails, but this is not
that important to fix.